[Improve][Connector-V2] Refactor the package of local file connector (#2403)
* [Improve][Connector-V2] Refactor the package of local file connector

* [Feature][Connector-V2] Add file system type in connector-file-local (#1946)
