+++ seatunnel-connectors-v2/connector-file/connector-file-local/src/main/java/org/apache/seatunnel/connectors/seatunnel/file/local/sink/filesystem/LocalFileSystem.java
@@ -1,0 +1,43 @@
+/*
+ * Licensed to the Apache Software Foundation (ASF) under one or more
+ * contributor license agreements.  See the NOTICE file distributed with
+ * this work for additional information regarding copyright ownership.
+ * The ASF licenses this file to You under the Apache License, Version 2.0
+ * (the "License"); you may not use this file except in compliance with
+ * the License.  You may obtain a copy of the License at
+ *
+ *    http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package org.apache.seatunnel.connectors.seatunnel.file.local.sink.filesystem;
+
+import org.apache.seatunnel.connectors.seatunnel.file.sink.spi.FileSystem;
+
+import java.io.File;
+import java.io.IOException;
+import java.util.Arrays;
+import java.util.List;
+
+public class LocalFileSystem implements FileSystem {
+    @Override
+    public void deleteFile(String path) throws IOException {
+        File file = new File(path);
+        file.delete();
+    }
+
+    @Override
+    public List<String> dirList(String dirPath) throws IOException {
+        File file = new File(dirPath);
+        String[] list = file.list();
+        if (list == null) {
+            return null;
+        }
+        return Arrays.asList(list);
+    }
+}
